frontend/requirements.txt
shiny
requests
Pillow